setting enviroment with maven and spring boot
