+++ src/main/java/com/agnolin/sanitarios/interfaces/GenericMapper.java
@@ -1,0 +1,5 @@
+package com.agnolin.sanitarios.interfaces;
+
+public interface GenericMapper {
+
+}
