+++ src/main/java/com/agnolin/sanitarios/interfaces/IService.java
@@ -1,0 +1,11 @@
+package com.agnolin.sanitarios.interfaces;
+
+import java.io.Serializable;
+
+public interface IService<T extends Serializable> {
+	
+	public void create(T entity);
+
+
+	public void update(T entity);
+}
