+++ src/main/java/com/agnolin/sanitarios/mappers/GenericMapperImple.java
@@ -1,0 +1,7 @@
+package com.agnolin.sanitarios.mappers;
+
+import com.agnolin.sanitarios.interfaces.GenericMapper;
+
+public class GenericMapperImple implements GenericMapper{
+
+}
