+++ src/main/java/com/agnolin/sanitarios/mappers/ProductMapper.java
@@ -1,0 +1,5 @@
+package com.agnolin.sanitarios.mappers;
+
+public class ProductMapper extends GenericMapperImple{
+
+}
