+++ src/main/java/com/agnolin/sanitarios/mappers/SaleDetailMapper.java
@@ -1,0 +1,5 @@
+package com.agnolin.sanitarios.mappers;
+
+public class SaleDetailMapper extends GenericMapperImple{
+
+}
