+++ src/main/java/com/agnolin/sanitarios/mappers/Salemapper.java
@@ -1,0 +1,5 @@
+package com.agnolin.sanitarios.mappers;
+
+public class Salemapper extends GenericMapperImple{
+
+}
