+++ src/main/java/com/agnolin/sanitarios/mappers/UserMapper.java
@@ -1,0 +1,5 @@
+package com.agnolin.sanitarios.mappers;
+
+public class UserMapper extends GenericMapperImple{
+
+}
