+++ src/main/java/com/agnolin/sanitarios/repository/GenericHibernateDao.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 @Scope(BeanDefinition.SCOPE_PROTOTYPE)
-public class GenericHibernateDao<T extends Serializable> extends AbstractHibernateDao<T> implements com.agnolin.sanitrarios.interfaces.IRepository<T> {
+public class GenericHibernateDao<T extends Serializable> extends AbstractHibernateDao<T> implements com.agnolin.sanitarios.interfaces.IRepository<T> {
