--- src/main/java/com/agnolin/sanitios/mappers/CategoriesMapper.java
@@ -1,5 +1,0 @@
-package com.agnolin.sanitios.mappers;
-
-public class CategoriesMapper {
-
-}
