--- src/main/java/com/agnolin/sanitios/mappers/GenericMapperImple.java
@@ -1,7 +1,0 @@
-package com.agnolin.sanitios.mappers;
-
-import com.agnolin.sanitrarios.interfaces.GenericMapper;
-
-public class GenericMapperImple implements GenericMapper{
-
-}
