--- src/main/java/com/agnolin/sanitios/mappers/Salemapper.java
@@ -1,5 +1,0 @@
-package com.agnolin.sanitios.mappers;
-
-public class Salemapper extends GenericMapperImple{
-
-}
