--- src/main/java/com/agnolin/sanitios/mappers/UserMapper.java
@@ -1,5 +1,0 @@
-package com.agnolin.sanitios.mappers;
-
-public class UserMapper extends GenericMapperImple{
-
-}
