--- src/main/java/com/agnolin/sanitrarios/interfaces/GenericMapper.java
@@ -1,5 +1,0 @@
-package com.agnolin.sanitrarios.interfaces;
-
-public interface GenericMapper {
-
-}
