--- src/main/java/com/agnolin/sanitrarios/interfaces/IService.java
@@ -1,11 +1,0 @@
-package com.agnolin.sanitrarios.interfaces;
-
-import java.io.Serializable;
-
-public interface IService<T extends Serializable> {
-	
-	public void create(T entity);
-
-
-	public void update(T entity);
-}
